Zip writer now supports concurrent access. (#1)
Calls to Create() will return a new FileHeader struct which will write
compressed data to a tmpfile. Upon Close() the data will be copied
again to the body of the zip file (while holding the mutex).

This allows multiple Zip writers to write a lot of data concurrently
and amortizes the compression cost over multiple cores.